neuro/requirements.txt
absl-py==2.2.2
actionlib==1.14.1
angles==1.9.14
app-manager==1.3.0
arbotix-python==0.11.0
arbotix_sensors==0.11.0
base_local_planner==1.17.3
bondpy==1.8.6
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.11
catkin-pkg==1.0.0
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.4
filelock==3.16.1
fingertip_pressure==1.9.0
fonttools==4.57.0
fsspec==2025.3.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gazebo_video_monitor_utils==0.7.1
gencpp==0.7.1
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.17
google-auth==2.39.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.10
image-geometry==1.16.2
image_cb_detector==0.10.15
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.1
Jinja2==3.1.6
joblib==1.4.2
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
joint_trajectory_action_tools==0.0.12
kdl-parser-py==1.14.2
kiwisolver==1.4.7
kobuki_dashboard==0.5.7
kobuki_qtestsuite==0.5.7
kobuki_testsuite==0.7.6
laser-geometry==1.6.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.17.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
mpmath==1.3.0
networkx==3.1
numpy==1.20.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
openni2_launch==1.6.1
packaging==25.0
pandas==2.0.3
Pillow==9.4.0
pkg_resources==0.0.0
pr2-tuck-arms-action==0.0.12
pr2_camera_synchronizer==1.6.32
pr2_computer_monitor==1.6.32
pr2_controller_manager==1.8.21
pr2_counterbalance_check==1.0.15
pr2_power_board==1.1.10
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.22.0
pycryptodomex==3.22.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-gnupg==0.5.4
python-qt-binding==0.4.5
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==0.4.3
qt-gui==0.4.3
qt-gui-cpp==0.4.3
qt-gui-py-common==0.4.3
requests==2.32.3
requests-oauthlib==2.0.0
resource-retriever==1.12.9
rosbag==1.17.0
rosboost-cfg==1.15.9
rosclean==1.15.9
roscreate==1.15.9
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.9
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.9
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospy==1.17.0
rosserial_python==0.9.2
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.9
roswtf==1.17.0
rqt-action==0.4.10
rqt-bag==0.5.2
rqt-bag-plugins==0.5.2
rqt-console==0.4.13
rqt-dep==0.4.13
rqt-graph==0.4.15
rqt-gui==0.5.4
rqt-gui-py==0.5.4
rqt-image-view==0.4.18
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.11
rqt-plot==0.4.15
rqt-pose-view==0.5.12
rqt-publisher==0.4.11
rqt-py-common==0.5.4
rqt-py-console==0.4.11
rqt-reconfigure==0.5.6
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.13
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.1
rqt-service-caller==0.4.11
rqt-srv==0.4.10
rqt-tf-tree==0.6.4
rqt-top==0.4.11
rqt-topic==0.4.14
rqt-web==0.4.11
rqt_launch==0.4.9
rqt_nav_view==0.5.7
rqt_shell==0.4.11
rsa==4.9.1
rviz==1.14.25
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sensor-msgs==1.13.1
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.6
sound-play==0.3.18
srdfdom==0.6.4
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tf==1.13.3
tf-conversions==1.13.3
tf2-geometry-msgs==0.7.8
tf2-kdl==0.7.8
tf2-py==0.7.8
tf2-ros==0.7.8
tf2-sensor-msgs==0.7.8
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
topic-tools==1.17.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
trac_ik_python==1.6.6
triton==3.0.0
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing_extensions==4.13.2
tzdata==2025.2
ur_driver==1.2.5
ur_kinematics==1.2.5
urdfdom-py==0.4.6
urllib3==2.2.3
Werkzeug==3.0.6
xacro==1.14.20
yocs_localization_manager==0.8.2
yocs_navigator==0.8.2
zipp==3.20.2
